Support outputing EventText as byte array field
Support outputing EventText as byte array field (i.e. binary field) when posting JSON to the remote service.

By default, EventText is output as JSON string field and the value of
EventText will be in UTF-8 encoded string. If EventText is output as
JSON byte array field, the value of EventText will be in base64-encoded string.